UDH_back/requirements.txt
asgiref==3.7.2
Django==4.2.10
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4

PyJWT~=2.8.0